removing duplicate implementation
